requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
supabase==2.23.0
starlette==0.37.2
uvicorn==0.29.0
pytz==2024.1